+++ CHANGELOG.md
@@ -8,12 +8,15 @@
 ## Changes
 
 - Support i18n(internationalization)
-    - Detect user system locate load different language packs
+    - Detect user system locate load different language packs in first time
+    - User can switch lanugages from Menu Items
+    - Save Language to settings.txt  
 
 ## Fix
 
 - Fix User Manual.pdf can not open in Mac & Linux
+- Disable Native Menu, Fix Mac Menu issue
 
 ## To-Do
 
-- Switch Language from Menu item
+- Switch Language from Menu item & Restart Program